Generate and reuse benchmark fixture across tests
